refactor: change the file location
