+++ src/data/footer/data-resources.ts
