+++ src/pages/_app.tsx
@@ -2,9 +2,9 @@ import { Header } from "../components/header";
 import "../styles/globals.css";
 import "reactflow/dist/style.css";
 import type { AppProps } from "next/app";
-import { DefaultSeo } from "next-seo"
-import SEO from "../../next-seo.config"
-import Footer from "../components/footer";
+import { DefaultSeo } from "next-seo";
+import SEO from "../../next-seo.config";
+import Footer from "../data/footer";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
